==> reasearch/01_data_ingetion.ipynb <==
import os

%pwd

os.chdir('../')

%pwd

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataIngestionConfig:
    root_dir: Path
    source_URL: str
    local_data_file: Path
    unzip_dir: Path

from src.mlproject1.constants import *
from src.mlproject1.utils.common import create_directories,read_yaml

class ConfigurationManager:
    def __init__(self,config_file_path=CONFIG_FILE_PATH,schema_file_path=SCHEMA_FILE_PATH,params_file_path=PARAMS_FILE_PATH):
        self.config=read_yaml(config_file_path)
        self.params=read_yaml(params_file_path)
        self.schema=read_yaml(schema_file_path)

        create_directories([self.config.artifacts_root])


    def get_data_ingestion_config(self) -> DataIngestionConfig:
        create_directories([self.config.data_ingetion.root_dir])

        data_ingestion_config=DataIngestionConfig(
            root_dir=self.config.data_ingetion.root_dir,
            source_URL=self.config.data_ingetion.source_URL,
            local_data_file=self.config.data_ingetion.local_data_file,
            unzip_dir=self.config.data_ingetion.unzip_dir
        )

        return data_ingestion_config
        
            

import os
from src.mlproject1 import logger
import urllib.request as request
import zipfile


class DataIngestion:
    def __init__(self,config: DataIngestionConfig):
        self.config=config

    def download_data(self):
        if not os.path.exists(self.config.data_ingetion.local_data_file):
            filename,header=request(
                url=self.config.data_ingetion.source_URL,
                filename=self.config.data_ingetion.local_data_file
            )
            logger.info(f"{filename} downloadad")

        else:
            logger.info(f"File already exists")

    def extract_zip_file(self):
        unzip_path=self.config.data_ingetion.unzip_dir
        os.makedirs(unzip_path,exist_ok=True)

        with zipfile.ZipFile(self.config.data_ingetion.local_data_file,'r') as zipf:
            zipf.extractall(unzip_path)

try:
    conf=ConfigurationManager()
    data_ingest=conf.get_data_ingestion_config()
    data_in=DataIngestion(data_ingest)
    data_in.download_data()
    data_in.extract_zip_file()
except Exception as e:
    raise e
