==> reasearch/02_data_validation.ipynb <==
import os

%pwd

os.chdir('../')

%pwd

import pandas as pd


from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    unzip_data_dir: Path
    STATUS_FILE: str
    all_schema: dict


from src.mlproject1.constants import *
from src.mlproject1.utils.common import read_yaml,create_directories
import os
from src.mlproject1 import logger

class ConfigurationManager:
    def __init__(self,CONFIG_FILE_PATH,SCHEMA_FILE_PATH,PARAMS_FILE_PATH):
        self.config=read_yaml(CONFIG_FILE_PATH)
        self.params=read_yaml(PARAMS_FILE_PATH)
        self.schema=read_yaml(SCHEMA_FILE_PATH)

        create_directories([self.config.artifacts_root])

    def get_data_validation_config(self) -> DataValidationConfig:
        try:
            config=self.config.data_validation
            schema=self.schema.COLUMNS

            create_directories([config.root_dir])

            data_validation_config= DataValidationConfig(
                root_dir=config.root_dir,
                unzip_data_dir=config.unzip_data_dir,
                STATUS_FILE=config.STATUS_FILE,
                all_schema=schema
            )

            return data_validation_config
        except Exception as e:
            raise e

import os
from src.mlproject1 import logger

class DataValidation:
    def __init__(self,config:DataValidationConfig):
        self.config=config

    def validate_all_columns(self)-> bool:
        try:
            validation_status=None
            data =pd.read_csv(self.config.unzip_data_dir)
            all_cols=list(data.columns)

            all_schema=self.config.all_schema.keys()

            for i in all_cols:
                if i not in all_schema:
                    validation_status= False
                    with open(self.config.STATUS_FILE,'w') as f:
                        f.write(f"Validation Status: {validation_status}")
                else:
                    validation_status= True
                    with open(self.config.STATUS_FILE,'w') as f:
                        f.write(f"Validation Status : {validation_status}")

            return validation_status

        except Exception as e:
            raise e


try:
    con=ConfigurationManager()
    data_val=con.get_data_validation_config()
    data_v=DataValidation(config=data_val)
    data_v.validate_all_columns()
except Exception as e:
    raise e
