==> reasearch/03_data_transformation.ipynb <==
%pwd

import os
os.chdir('../')

%pwd

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataTransformationConfig:
    root_dir: Path
    data_path: Path

from src.mlproject1.constants import *
from src.mlproject1.utils.common import create_directories,read_yaml

class ConfigurationManager():
    def __init__(self,CONFIG_FILE_PATH,SCHEMA_FILE_PATH,PARAMS_FILE_PATH):
        self.config=read_yaml(CONFIG_FILE_PATH)
        self.params=read_yaml(PARAMS_FILE_PATH)
        self.schema=read_yaml(SCHEMA_FILE_PATH)

        create_directories([self.config.root_dir])

    def get_data_transformation_config(self)-> DataTransformationConfig:

        config=self.config.data_transformation

        create_directories([config.root_dir])
        create_directories([config.data_path])

        data_transformation_config=DataTransformationConfig(
            root_dir=config.root_dir,
            data_path=config.data_path
        )

        return data_transformation_config

import pandas as pd
from sklearn.model_selection import train_test_split

class DataTransformation:
    def __init__(self,config: DataTransformationConfig):
        self.config=config

    def train_test_split(self):
        data=pd.read_csv(self.config.data_path)

        train,test=train_test_split(data)

        train.to_csv(os.path.join(self.config.root_dir,"train.csv"),index= False)
        test.to_csv(os.path.join(self.config.root_dir,"test.csv"),index= False)

try:
    conf=ConfigurationManager()
    data_trans=conf.get_data_transformation_config()
    data_t=DataTransformation(config=data_trans)
    data_t.train_test_split
except Exception as e:
    raise e
